requirements.txt
Flask==1.0.2
line-bot-sdk==1.8.0
psycopg2==2.8.2
